requirements.txt
Flask==2.1.1
openai==0.27.0
tweepy==4.3.0
python-dotenv==0.19.2
apscheduler==3.8.1
Pillow==8.3.2